disable the RGB API by default
