+++ fuzz/cgif_rgb_create_fuzz_seed.c
@@ -4,6 +4,7 @@
 */
 
 #include <cgif.h>
+#include <cgif_rgb.h>
 
 #include <stdio.h>
 #include <stdint.h>
