+++ inc/cgif.h
@@ -26,9 +26,6 @@ extern "C" {
 
 #define CGIF_INFINITE_LOOP               (0x0000uL)       // for animated GIF: 0 specifies infinite loop
 
-#define CGIF_RGB_FRAME_ATTR_INTERLACED   (1ul << 0)       // encode frame interlaced (default is not interlaced)
-#define CGIF_RGB_FRAME_ATTR_NO_DITHERING (1ul << 1)       // disable color dithering (default is with dithering)
-
 typedef enum {
   CGIF_ERROR = -1, // something unspecified failed
   CGIF_OK    =  0, // everything OK
@@ -41,17 +38,9 @@ typedef enum {
   CGIF_PENDING,
 } cgif_result;
 
-typedef enum {
-  CGIF_CHAN_FMT_RGB  = 3, // 3 byte per pixel (red, green, blue)
-  CGIF_CHAN_FMT_RGBA = 4, // 4 byte per pixel (red, green, blue, alpha)
-} cgif_chan_fmt;
-
 typedef struct st_gif                  CGIF;              // struct for the full GIF
 typedef struct st_gifconfig            CGIF_Config;       // global cofinguration parameters of the GIF
 typedef struct st_frameconfig          CGIF_FrameConfig;  // local configuration parameters for a frame
-typedef struct st_cgif_rgb_config      CGIFrgb_Config;
-typedef struct st_cgif_rgb             CGIFrgb;
-typedef struct st_cgif_rgb_frameconfig CGIFrgb_FrameConfig;
 
 typedef int cgif_write_fn(void* pContext, const uint8_t* pData, const size_t numBytes); // callback function for stream-based output
 
@@ -59,10 +48,6 @@ typedef int cgif_write_fn(void* pContext, const uint8_t* pData, const size_t num
 CGIF* cgif_newgif     (CGIF_Config* pConfig);                  // creates a new GIF (returns pointer to new GIF or NULL on error)
 int   cgif_addframe   (CGIF* pGIF, CGIF_FrameConfig* pConfig); // adds the next frame to an existing GIF (returns 0 on success)
 int   cgif_close      (CGIF* pGIF);                          // close file and free allocated memory (returns 0 on success)
-
-CGIFrgb*    cgif_rgb_newgif    (const CGIFrgb_Config* pConfig);
-cgif_result cgif_rgb_addframe  (CGIFrgb* pGIF, const CGIFrgb_FrameConfig* pConfig);
-cgif_result cgif_rgb_close     (CGIFrgb* pGIF);
 
 // CGIF_Config type (parameters passed by user)
 // note: must stay AS IS for backward compatibility
@@ -89,25 +74,6 @@ struct st_frameconfig {
   uint16_t  delay;                                     // delay before the next frame is shown (units of 0.01 s)
   uint16_t  numLocalPaletteEntries;                    // size of the local color table
   uint8_t   transIndex;                                // introduced with V0.2.0
-};
-
-struct st_cgif_rgb_config {
-  cgif_write_fn* pWriteFn;
-  void*          pContext;
-  const char*    path;
-  uint32_t       attrFlags;
-  uint32_t       genFlags;
-  uint16_t       numLoops;
-  uint16_t       width;
-  uint16_t       height;
-};
-
-struct st_cgif_rgb_frameconfig {
-  uint8_t* pImageData;
-  cgif_chan_fmt fmtChan;
-  uint32_t attrFlags;   // TBD
-  uint32_t genFlags;    // TBD
-  uint16_t delay;
 };
 
 #ifdef __cplusplus
