+++ tests/rgb_256digit.c
@@ -4,6 +4,7 @@
 #include <stdio.h>
 
 #include "cgif.h"
+#include "cgif_rgb.h"
 
 #define WIDTH  1000
 #define HEIGHT 1000
