+++ tests/rgb_interlaced.c
@@ -4,6 +4,7 @@
 #include <stdio.h>
 
 #include "cgif.h"
+#include "cgif_rgb.h"
 
 #define WIDTH 50
 #define HEIGHT 50
