+++ tests/rgb_single_color.c
@@ -4,6 +4,7 @@
 #include <stdio.h>
 
 #include "cgif.h"
+#include "cgif_rgb.h"
 
 #define WIDTH  200
 #define HEIGHT 200
